move the test to misc_test
